+++ lib/stenotype/event.rb
@@ -11,19 +11,19 @@ module Stenotype
     # @example Emit an event using class method
     #   Stenotype::Event.emit!(data, options, eval_context)
     #
-    # @param data {Hash} Data to be published to the targets.
-    # @param options {Hash} A hash of additional options to be tracked.
-    # @param eval_context {Hash} A context having handler defined in {Stenotype::ContextHandlers}.
-    # @param dispatcher {#publish} A dispatcher object responding to [#publish]
+    # @param {[String, Symbol]} name Event name.
+    # @param {Hash} attributes Data to be published to the targets.
+    # @param {Hash} eval_context A context having handler defined in {Stenotype::ContextHandlers}.
+    # @param dispatcher {#publish} A dispatcher object responding to [#publish].
     # @return {Stenotype::Event} An instance of {Stenotype::Event}
     #
-    def self.emit!(data, options: {}, eval_context: {}, dispatcher: Stenotype.config.dispatcher)
-      event = new(data, options: options, eval_context: eval_context, dispatcher: dispatcher)
+    def self.emit!(name, attributes = {}, eval_context: {}, dispatcher: Stenotype.config.dispatcher)
+      event = new(name, attributes, eval_context: eval_context, dispatcher: dispatcher)
       event.emit!
       event
     end
 
-    attr_reader :data, :options, :eval_context, :dispatcher
+    attr_reader :name, :attributes, :eval_context, :dispatcher
 
     #
     # @example Create an event
@@ -31,15 +31,15 @@ module Stenotype
     # @example Create an event with custom dispatcher
     #   event = Stenotype::Event.new(data, options, eval_context, dispatcher: MyDispatcher.new)
     #
-    # @param {Hash} data Data to be published to the targets.
-    # @param {Hash} options A hash of additional options to be tracked.
+    # @param {[String, Symbol]} name Event name.
+    # @param {Hash} attributes Data to be published to the targets.
     # @param {Hash} eval_context A context having handler defined in {Stenotype::ContextHandlers}.
     # @param dispatcher {#publish} A dispatcher object responding to [#publish].
     # @return {Stenotype::Event} An instance of event
     #
-    def initialize(data, options: {}, eval_context: {}, dispatcher: Stenotype.config.dispatcher)
-      @data = data
-      @options = options
+    def initialize(name, attributes = {}, eval_context: {}, dispatcher: Stenotype.config.dispatcher)
+      @name = name
+      @attributes = attributes
       @eval_context = eval_context
       @dispatcher = dispatcher.new
     end
@@ -48,7 +48,7 @@ module Stenotype
     # Emits a {Stenotype::Event}.
     #
     # @example Emit an instance of event
-    #   event = Stenotype::Event.new(data, options, eval_context)
+    #   event = Stenotype::Event.new('events_name', { key: :value }, eval_context: { controller: ctrl })
     #   event.emit! #=> Publishes the event to targets
     #
     def emit!