+++ spec/stenotype/event_serializer_spec.rb
@@ -5,14 +5,14 @@ require "spec_helper"
 RSpec.describe Stenotype::EventSerializer do
   subject(:serializer) { described_class.new(event, uuid_generator: Stenotype::TestUuidGen) }
 
+  let(:event_name) { "event_name" }
   let(:data) { { data_key: "data value" } }
-  let(:options) { { options_key: "options value" } }
   let(:eval_context) { { klass: dummy_context.new } }
 
   let(:event) do
     Stenotype::Event.new(
+      event_name,
       data,
-      options: options,
       eval_context: eval_context,
     )
   end
@@ -22,8 +22,8 @@ RSpec.describe Stenotype::EventSerializer do
   describe "#serialize", type: :with_frozen_time do
     it "represents an event as a hash" do
       expect(serializer.serialize).to eq(
-        data_key: "data value",
-        options_key: "options value",
+        name: event_name,
+        **data,
         timestamp: Time.now.utc,
         uuid: "abcd",
       )